this update fixes package-lock and also fixes npm stuff, also fixes game images.
**IMPORTANT**

- game images fixed

** BUG FIXES **

- images was NOT showing due to no images matching the names

** WHAT ELSE? **

i modified .gitignore to remove a script which was crucial to me as a developer.

** NEXT UPDATE PLANS?!?!! **

maybe ill add another few games, and also im going to add descriptions at the bottom of the titles inside the games so it will show how to play and what its about.

** small stuff **

changed the layout making some games bigger depending how popular they are.
edited readme to look nice

# NOT ALL GAMES IMAGES ARE IMPORTED!